requirements.txt
Flask==3.0.3
paramiko==3.4.0
azure-storage-blob==12.19.1
lxml==5.2.1
Werkzeug==3.0.3
gunicorn==22.0.0
